+++ test/unit/elastic_db_admin/run_program.py
@@ -35,32 +35,69 @@ import version
 __version__ = version.__version__
 
 
-def failed_dumps(es, **kwargs):
+def failed_dumps(els, **kwargs):
 
     """Function:  failed_dumps
 
     Description:  This is a function stub for elastic_db_admin.failed_dumps.
 
     Arguments:
-        es -> Stub argument holder.
+        els -> Stub argument holder.
 
     """
 
-    pass
+    status = True
+
+    if els and dict(kwargs.get("args_array")):
+        status = True
+
+    return status
 
 
-def list_dumps(es, **kwargs):
+def list_dumps(els, **kwargs):
 
     """Function:  list_repos
 
     Description:  This is a function stub for elastic_db_admin.list_dumps.
 
     Arguments:
-        es -> Stub argument holder.
+        els -> Stub argument holder.
 
     """
 
-    pass
+    status = True
+
+    if els and dict(kwargs.get("args_array")):
+        status = True
+
+    return status
+
+
+class ProgramLock(object):
+
+    """Class:  ProgramLock
+
+    Description:  Class stub holder for gen_class.ProgramLock class.
+
+    Methods:
+        __init__ -> Class initialization.
+
+    """
+
+    def __init__(self, cmdline, flavor):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+            (input) cmdline -> Argv command line.
+            (input) flavor -> Lock flavor ID.
+
+        """
+
+        self.cmdline = cmdline
+        self.flavor = flavor
 
 
 class UnitTest(unittest.TestCase):
@@ -71,6 +108,7 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp -> Unit testing initilization.
+        test_raise_exception -> Test with raising exception.
         test_func_call_multi -> Test run_program with multiple calls.
         test_func_call_one -> Test run_program with one call to function.
         test_func_call_zero -> Test run_program with zero calls to function.
@@ -111,13 +149,14 @@ class UnitTest(unittest.TestCase):
                 self.host = ["SERVER_NAME"]
                 self.port = 9200
 
-        self.ct = CfgTest()
+        self.cfg = CfgTest()
 
         self.args = {"-c": "config_file", "-d": "config_dir", "-M": True}
         self.func_dict = {"-F": failed_dumps, "-L": list_dumps}
+        self.proglock = ProgramLock(["cmdline"], "FlavorID")
 
     @mock.patch("elastic_db_admin.gen_libs.load_module")
-    @mock.patch("elastic_db_admin.elastic_class.ElasticSearch")
+    @mock.patch("elastic_db_admin.elastic_class.ElasticSearchStatus")
     @mock.patch("elastic_db_admin.gen_class.ProgramLock")
     def test_raise_exception(self, mock_lock, mock_class, mock_load):
 
@@ -134,14 +173,14 @@ class UnitTest(unittest.TestCase):
         mock_lock.side_effect = \
             elastic_db_admin.gen_class.SingleInstanceException
         mock_class.return_value = "ElasticSearch"
-        mock_load.return_value = self.ct
+        mock_load.return_value = self.cfg
 
         with gen_libs.no_std_out():
             self.assertFalse(elastic_db_admin.run_program(self.args,
                                                           self.func_dict))
 
     @mock.patch("elastic_db_admin.gen_libs.load_module")
-    @mock.patch("elastic_db_admin.elastic_class.ElasticSearch")
+    @mock.patch("elastic_db_admin.elastic_class.ElasticSearchStatus")
     @mock.patch("elastic_db_admin.gen_class")
     def test_func_call_multi(self, mock_lock, mock_class, mock_load):
 
@@ -157,15 +196,15 @@ class UnitTest(unittest.TestCase):
         self.args["-F"] = True
         self.args["-L"] = True
 
-        mock_lock.ProgramLock = elastic_db_admin.gen_class.ProgramLock
+        mock_lock.return_value = self.proglock
         mock_class.return_value = "ElasticSearch"
-        mock_load.return_value = self.ct
+        mock_load.return_value = self.cfg
 
         self.assertFalse(elastic_db_admin.run_program(self.args,
                                                       self.func_dict))
 
     @mock.patch("elastic_db_admin.gen_libs.load_module")
-    @mock.patch("elastic_db_admin.elastic_class.ElasticSearch")
+    @mock.patch("elastic_db_admin.elastic_class.ElasticSearchStatus")
     @mock.patch("elastic_db_admin.gen_class")
     def test_func_call_one(self, mock_lock, mock_class, mock_load):
 
@@ -179,15 +218,15 @@ class UnitTest(unittest.TestCase):
 
         self.args["-F"] = True
 
-        mock_lock.ProgramLock = elastic_db_admin.gen_class.ProgramLock
+        mock_lock.return_value = self.proglock
         mock_class.return_value = "ElasticSearch"
-        mock_load.return_value = self.ct
+        mock_load.return_value = self.cfg
 
         self.assertFalse(elastic_db_admin.run_program(self.args,
                                                       self.func_dict))
 
     @mock.patch("elastic_db_admin.gen_libs.load_module")
-    @mock.patch("elastic_db_admin.elastic_class.ElasticSearch")
+    @mock.patch("elastic_db_admin.elastic_class.ElasticSearchStatus")
     @mock.patch("elastic_db_admin.gen_class")
     def test_func_call_zero(self, mock_lock, mock_class, mock_load):
 
@@ -199,9 +238,9 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        mock_lock.ProgramLock = elastic_db_admin.gen_class.ProgramLock
+        mock_lock.return_value = self.proglock
         mock_class.return_value = "ElasticSearch"
-        mock_load.return_value = self.ct
+        mock_load.return_value = self.cfg
 
         self.assertFalse(elastic_db_admin.run_program(self.args,
                                                       self.func_dict))